requirements.txt
asyncio
ffmpeg
hachoir
pyrofork==2.3.61
python-dotenv
pytz
TgCrypto
psutil
